requirements.txt
dotenv==0.9.9
mysql-connector-python==9.6.0
numpy==2.4.4
pandas==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
tzdata==2026.2
